+++ docs/examples/example_meshing_dual.py
@@ -1,20 +1,24 @@
 from pathlib import Path
+from compas.geometry import Line
 from compas.datastructures import Mesh
 from compas_viewer import Viewer
 
-from compas_cgal.meshing import mesh_remesh
+from compas_cgal.meshing import remesh_dual
 
 
 input_file = Path(__file__).parent.parent.parent / "data" / "rhinovault_mesh_0.ply"
 mesh = Mesh.from_ply(input_file)
 mesh.quads_to_triangles()
 
-edge_length = 1
-iterations = 10
 
-# Remesh
-V1, F1 = mesh_remesh(mesh.to_vertices_and_faces(), edge_length, iterations)
-remeshed = Mesh.from_vertices_and_faces(V1, F1)
+fixed_vertices = []
+for v in mesh.vertices():
+    if mesh.vertex_point(v).z < 0.01:
+        fixed_vertices.append(v)
+
+# Remesh and Dual
+V2, F2, V3, F3 = remesh_dual(mesh.to_vertices_and_faces(), length_factor=1.0, number_of_iterations=100, angle_radians=0.9, fixed_vertices=[])
+remeshed, dual = Mesh.from_vertices_and_faces(V2, F2), Mesh.from_vertices_and_faces(V3, F3)
 
 
 # ==============================================================================
@@ -27,6 +31,11 @@ viewer.renderer.camera.target = [0, 0, 0]
 viewer.renderer.camera.position = [0, -0.25, 0.10]
 
 viewer.scene.add(mesh, show_points=False, opacity=0.25)
+remeshed.translate([0, 0, 3])
 viewer.scene.add(remeshed, show_points=True)
+viewer.scene.add(dual, show_points=True)
+
+for v in dual.vertices():
+    viewer.scene.add(Line(dual.vertex_point(v), dual.vertex_point(v)+dual.vertex_normal(v)*0.25), color=(255, 0, 0))
 
 viewer.show()